fix: unblock main CI — dead-stripped symbol + cargo-test staticlib + oversized link/mod.rs
Three pre-existing CI fragilities on main turning required checks red for PRs:

0. cargo-test runs 'cargo test -p perry-runtime', which never builds the
   staticlib crate-type, so libperry_runtime.a/libperry_stdlib.a only exist from
   cache. A PR touching perry-runtime invalidates the cached staticlib and cargo
   never rebuilds it -> PERRY_NO_AUTO_OPTIMIZE=1 tests fail with 'Could not find
   libperry_runtime.a'. Fix: add 'cargo build -p perry-runtime -p perry-stdlib'
   to the cargo-test job before the integration-test loop.

1. js_array_numeric_value_to_raw_f64 (#5291) is #[no_mangle] but called only
   from generated code, so the linker dead-strips it from libperry_runtime.a ->
   undefined symbol at link. Fix: #[used] keepalive anchor.

2. link/mod.rs crossed the 2000-line file-size gate (2132) after #5400. Split
   the ~1770-line build_and_run_link orchestrator into link/build_and_run.rs.
